[ksqlDb.RestApi.Client]: model builder example
