--- Samples/ksqlDB.RestApi.Client.Sample/Model/PaymentModelBuilder.cs
@@ -1,43 +1,0 @@
-using ksqlDb.RestApi.Client.FluentAPI.Builders;
-using ksqlDB.RestApi.Client.KSql.RestApi;
-using ksqlDB.RestApi.Client.KSql.RestApi.Http;
-
-namespace ksqlDB.RestApi.Client.Samples.Model
-{
-  public class PaymentModelBuilder
-  {
-    private record Payment
-    {
-      public string Id { get; set; } = null!;
-      public decimal Amount { get; set; }
-      public string Description { get; set; } = null!;
-    }
-    private record Account
-    {
-      public string Id { get; set; } = null!;
-      public decimal Balance { get; set; }
-    }
-
-    private static void InitModel()
-    {
-      ModelBuilder builder = new();
-
-      builder.Entity<Account>()
-        .HasKey(c => c.Id)
-        .Property(b => b.Balance);
-
-      builder.Entity<Payment>()
-        .HasKey(c => c.Id)
-        .Property(b => b.Amount)
-        .Decimal(precision: 10, scale: 2);
-
-      var httpClient = new HttpClient
-      {
-        BaseAddress = new Uri("http://localhost:8088")
-      };
-      var httpClientFactory = new HttpClientFactory(httpClient);
-
-      var restApiProvider = new KSqlDbRestApiClient(httpClientFactory);
-    }
-  }
-}
